Address third review round: enforce the dependency rule, complete the file tree
Stop app tests from importing model packages. apps/crazy_robotaxi/tests/
now builds a stub pipeline config from base flashdreams config classes;
the tests that genuinely needed OmniDreams internals moved to
integrations_v2/omnidreams/tests/test_recipe_configs.py, beside the
interactive_drive adapter tests already there. No app test imports a
model package now.

Nest apps/v2v/ to apps/v2v/v2v/ with a sibling tests/, replacing its
package-dir remap with the same auto-discovery every other app uses. It
was the last app not matching the documented shape.

Fix claims in AGENTS.md that were not true: apps do import
integrations_v2 in one live-feature path, and the submodule list omitted
runtime_v2, the most-used one. Rename Boundaries to Dependencies and
state every direction the same way. Generalize the .claude/worktrees
note to gitignored AI-tool directories.

Complete the CONTRIBUTING.md file tree with the subpackages it omitted
(runtime/, serving/, demo/, accelerated/, quality/, _pytest_plugins/),
one entry per line, and note that the five fixture packages do not
follow the model shape.

Drop remaining references to the deleted v1 integrations/ directory from
README.md, docs/README.md, and docs/source/, including a link that 404s
and two sentences that only existed to contrast against v1.
